Shows standard deviation and medium on graphs. Closes #68

+++ gatling-statistics/src/main/scala/com/excilys/ebi/gatling/statistics/extractor/DetailsRequestsDataExtractor.scala
@@ -14,8 +14,8 @@ class DetailsRequestsDataExtractor(val runOn: String) extends Logging {
     computeStatistics(extractedInformation)
   }
 
-  private def extractFromFile: Map[String, List[(String, Int)]] = {
-    var extractedValues: Map[String, List[(String, Int)]] = TreeMap.empty
+  private def extractFromFile: Map[String, List[(String, Double)]] = {
+    var extractedValues: Map[String, List[(String, Double)]] = TreeMap.empty
     for (line <- Source.fromFile("results/" + runOn + "/simulation.log", "utf-8").getLines) {
       // Split each line by tabulation (As we get data from a TSV file)
       line.split("\t") match {
@@ -23,7 +23,7 @@ class DetailsRequestsDataExtractor(val runOn: String) extends Logging {
         case Array(runOn, scenarioName, userId, actionName, executionStartDate, executionDuration, resultStatus, resultMessage) => {
           if (actionName != "End of scenario")
             extractedValues = extractedValues + (actionName ->
-              ((executionStartDate, executionDuration.toInt) :: extractedValues.getOrElse(actionName, Nil)))
+              ((executionStartDate, executionDuration.toDouble) :: extractedValues.getOrElse(actionName, Nil)))
         }
         // Else, if the resulting data is not well formated print an error message
         case _ => sys.error("Input file not well formatted")
@@ -32,7 +32,7 @@ class DetailsRequestsDataExtractor(val runOn: String) extends Logging {
     extractedValues
   }
 
-  private def computeStatistics(extractedInformation: Map[String, List[(String, Int)]]): Map[String, DetailsRequestsDataResult] = {
+  private def computeStatistics(extractedInformation: Map[String, List[(String, Double)]]): Map[String, DetailsRequestsDataResult] = {
     var results: Map[String, DetailsRequestsDataResult] = Map.empty
     extractedInformation.foreach {
       case (requestName, responseTimes) =>
