+++ gatling-statistics/src/main/scala/com/excilys/ebi/gatling/statistics/template/Series.scala
@@ -1,6 +1,6 @@
 package com.excilys.ebi.gatling.statistics.template
 
-class Series(val name: String, val data: List[(String, Int)]) {
+class Series(val name: String, val data: List[(String, Double)]) {
   override def toString = {
     "name: '" + name + "', data: " + data.map { e => "[" + e._1 + ", " + e._2 + "]" }.mkString("[ ", ",", " ]")
   }
